latest - Still not working. Did some debugging, found out serial print messes with the data ready line

+++ EPC901_Reader/EPC901_Reader.h
@@ -2,7 +2,9 @@
 void flush();
 void flushBuffer();
 void capture(long);
-unsigned int readPixel();
+
+uint_least8_t readPixel();
+//unsigned int readPixel();
 bool readPicture();//int[]);
 void epcSleep();
 void epcWake();
